Added profile bar graph component
